+++ src/main/java/no/uib/inf101/sem2/gameEngine/model/shape/GridPosition.java
@@ -1,15 +1,15 @@
 package no.uib.inf101.sem2.gameEngine.model.shape;
 
 public interface GridPosition{
-    default public double x(){
-        return 0.0;
+    default public float x(){
+        return 0.0f;
     };
 
-    default public double y(){
-        return 0.0;
+    default public float y(){
+        return 0.0f;
     };
 
-    default public double z(){
-        return 0.0;
+    default public float z(){
+        return 0.0f;
     };
 }
