+++ src/main/java/no/uib/inf101/sem2/gameEngine/model/shape/Position2D.java
@@ -1,5 +1,5 @@
 package no.uib.inf101.sem2.gameEngine.model.shape;
 
-public record Position2D(double x, double y) implements GridPosition {
+public record Position2D(float x, float y) implements GridPosition {
     
 }
