+++ src/main/java/no/uib/inf101/sem2/gameEngine/model/shape/Position3D.java
@@ -1,5 +1,5 @@
 package no.uib.inf101.sem2.gameEngine.model.shape;
 
-public record Position3D(double x, double y, double z) implements GridPosition {
+public record Position3D(float x, float y, float z) implements GridPosition {
     
 }
