+++ src/main/java/no/uib/inf101/sem2/gameEngine/view/raycaster/LinearMath/ViewProjectionMatrix.java
@@ -9,17 +9,17 @@ public class ViewProjectionMatrix {
     Matrix projectionMatrix;
     Matrix viewProjectionMatrix;
 
-    public ViewProjectionMatrix(double fov, double aspectRatio, double near, double far, Rotation rot){
+    public ViewProjectionMatrix(float fov, float aspectRatio, float near, float far, Rotation rot){
         createProjectionMatrix(fov, aspectRatio, near, far);
         createViewMatrix(rot);
         createViewProjectionMatrix(this.viewMatrix, this.projectionMatrix);
     }
 
 
-    private void createProjectionMatrix(double fov, double aspectRatio, double near, double far){
-        double tanHalfFov = Math.tan(fov / 2); 
+    private void createProjectionMatrix(float fov, float aspectRatio, float near, float far){
+        float tanHalfFov = (float) Math.tan(fov / 2); 
 
-        Matrix pMatrix = new Matrix( new double[][] {
+        Matrix pMatrix = new Matrix( new float[][] {
             {1/(aspectRatio*tanHalfFov), 0           , 0                       , 0                     },
             {0                         , 1/tanHalfFov, 0                       , 0                     },
             {0                         , 0           , -(far+near)/(far-near)  , -2*far*near/(far-near)},
@@ -35,8 +35,8 @@ public class ViewProjectionMatrix {
     }
 
     private void createViewProjectionMatrix(Matrix viewM, Matrix ProjM){
-        double[][] m = viewM.value;
-        Matrix expandedViewMatrix = new Matrix(new double[][]{
+        float[][] m = viewM.value;
+        Matrix expandedViewMatrix = new Matrix(new float[][]{
             {m[0][0], m[0][1], m[0][2], 0},
             {m[1][0], m[1][1], m[1][2], 0},
             {m[2][0], m[2][1], m[2][2], 0},
@@ -54,13 +54,13 @@ public class ViewProjectionMatrix {
         if(v.getDims() != 3){
             throw new IllegalArgumentException("Vector must have 3 dimensions");
         }
-        Vector homogenousV = new Vector(new double[] {v.value[0], v.value[1], v.value[2], 1});
+        Vector homogenousV = new Vector(new float[] {v.value[0], v.value[1], v.value[2], 1});
         Vector transformedV = this.projectionMatrix.multiply(homogenousV);
 
-        double x = transformedV.value[0]/transformedV.value[3];
-        double y = transformedV.value[1]/transformedV.value[3];
+        float x = transformedV.value[0]/transformedV.value[3];
+        float y = transformedV.value[1]/transformedV.value[3];
         
-        return new Vector(new double[] {x, y});
+        return new Vector(new float[] {x, y});
 
     }
